requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
schedule>=1.2.0
websocket-client>=1.6.0
pydantic>=2.0.0
aiohttp>=3.8.0
